Extract common code from HTML and AMP `Text` exporters
AMP is basically just a custom version of HTML. Therefore, the code of
the two exporters is very similar. For better reusability (e.g. when
implementing a Facebook Instant Article exporter), extract the code into
a `Common` module which then is reused by both classes.
